Remove /overview path from insights
This will only be used for the dashboard, everything else will use /rules
