+++ src/SmartComponents/Rules/Details.scss
