+++ src/SmartComponents/Rules/ListRules.js
@@ -3,6 +3,7 @@ import routerParams from '@redhat-cloud-services/frontend-components-utilities/f
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
+import { PageHeader, PageHeaderTitle } from '@redhat-cloud-services/frontend-components';
 
 import * as AppActions from '../../AppActions';
 import RulesTable from '../../PresentationalComponents/RulesTable/RulesTable';
@@ -13,7 +14,11 @@ class ListRules extends Component {
         this.props.setBreadcrumbs([{ title: 'Rules', navigate: '/rules' }]);
     }
 
-    render = () => <RulesTable/>;
+    render = () => <>
+        <PageHeader>
+            <PageHeaderTitle title='Rules'/>
+        </PageHeader>
+        <RulesTable/></>;
 }
 
 ListRules.displayName = 'list-rules';
